move restoreServiceProviderLoggers to @sap-ux/fiori-generator-shared
