+++ .changeset/move-restore-service-provider-loggers-to-shared.md
@@ -1,0 +1,6 @@
+---
+'@sap-ux/fiori-generator-shared': patch
+'@sap-ux/fiori-app-sub-generator': patch
+---
+
+refactor: move restoreServiceProviderLoggers to @sap-ux/fiori-generator-shared
