+++ packages/fiori-app-sub-generator/src/utils/common.ts
@@ -1,5 +1,5 @@
 import { convert } from '@sap-ux/annotation-converter';
-import type { Annotations, ServiceProvider } from '@sap-ux/axios-extension';
+import type { Annotations } from '@sap-ux/axios-extension';
 import { isAbapEnvironmentOnBtp, isAppStudio } from '@sap-ux/btp-utils';
 import type { CapRuntime, CapService } from '@sap-ux/cap-config-writer';
 import { checkCdsUi5PluginEnabled, getAppLaunchText } from '@sap-ux/cap-config-writer';
@@ -316,27 +316,4 @@ export async function getAnnotations(
     }
 }
 
-/**
- * Restore the loggers for the service provider if they are missing.
- * This is necessary because the service provider may have been serialized and deserialized, which can lead to missing loggers which contain circular refs.
- * Not doing this will result in the loggers being undefined when trying to access them, and calling services will throw.
- *
- * @param logger - The logger instance to be restored.
- * @param serviceProvider - The service provider object that may have missing loggers.
- * @returns The service provider with restored loggers.
- */
-export function restoreServiceProviderLoggers(
-    logger: Logger,
-    serviceProvider?: ServiceProvider
-): ServiceProvider | undefined {
-    // Restore the loggers if missing.
-    for (const service in (serviceProvider as any)?.services) {
-        if ((serviceProvider as any).services?.[service].log && !(serviceProvider as any).services[service].log.info) {
-            (serviceProvider as any).services[service].log = logger;
-        }
-    }
-    if (serviceProvider?.log && !serviceProvider.log.info) {
-        serviceProvider.log = logger;
-    }
-    return serviceProvider;
-}
+export { restoreServiceProviderLoggers } from '@sap-ux/fiori-generator-shared';
